basic sentence structure and dictionary
Need special case for “I” (I am, I was), in other cases can treat as
plural

Should we write rules for noun? For most nouns, the plural form is just
adding “s” or “es” (with lots of exceptions of course). If it’s
possible to write a rule, then the dictionary part will be a lot
easier. Same for verb.